refactor: rename package to chatmd and add direnv setup
