+++ .gitignore
@@ -15,3 +15,4 @@ __pycache__
 
 # not project code
 data
+sinex_md/
